+++ internal/reconcile/vernemq.go
@@ -213,6 +213,10 @@ func getVerneMQEnvVars(statefulSetName string, cr *apiv1alpha2.Astarte) []v1.Env
 			Name:  "DOCKER_VERNEMQ_KUBERNETES_LABEL_SELECTOR",
 			Value: "app=" + statefulSetName,
 		},
+		{
+			Name:      "RELEASE_COOKIE",
+			ValueFrom: getErlangClusteringCookieSecretReference(cr),
+		},
 	}
 
 	if cr.Spec.AstarteInstanceID != "" {
@@ -467,7 +471,7 @@ func getVerneMQPolicyRules() []rbacv1.PolicyRule {
 	return []rbacv1.PolicyRule{
 		{
 			APIGroups: []string{""},
-			Resources: []string{"pods", "services"},
+			Resources: []string{"pods", "services", "endpoints"},
 			Verbs:     []string{"list", "get"},
 		},
 		{