+++ CHANGELOG.md
@@ -13,6 +13,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fix the propagation of replicaCount, installCrds, and pullPolicy values
   in values.yaml during Helm chart installation.
 
+## [24.5.2] - Unreleased
+### Added
+- AppEngine, DUP, VerneMQ have `get` and `list` permissions on
+  endpoints, pods and services.
+
+### Changed
+- AppEngine, DUP, VerneMQ use the same Erlang cookie.
+  This will not affect Astartes up to and including 1.2.0.
+
 ## [24.5.1] - 2025-01-10
 ### Fixed
 - Properly inject cassandra user credentials in the VerneMQ statefulset.